# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FlightRegistrationInput {
  deviceId: string;
  flightNumber: string;
  flightDate: string;
  passengerName?: string;
}

export interface FlightRegistrationResponse {
  id: number;
  deviceId: string;
  flightNumber: string;
  flightDate: string;
  /** @nullable */
  passengerName?: string | null;
  registeredAt: string;
}

export interface UnregisterFlightRequest {
  deviceId: string;
  flightNumber: string;
  flightDate: string;
}

export interface UnregisterResult {
  success: boolean;
}

export interface FlightPassengers {
  flightNumber: string;
  flightDate: string;
  count: number;
  names: string[];
}

export interface RatingInput {
  deviceId: string;
  flightNumber: string;
  flightDate: string;
  iataCode?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}

export interface RatingResult {
  success: boolean;
  totalRatings: number;
  /** @nullable */
  averageRating: number | null;
}

export interface FlightRatingAggregate {
  flightNumber: string;
  /** @nullable */
  flightDate: string | null;
  totalRatings: number;
  /** @nullable */
  averageRating: number | null;
}

export interface AirlineRatingAggregate {
  iataCode: string;
  totalRatings: number;
  /** @nullable */
  averageRating: number | null;
}

export interface AirlineRatingSummaryItem {
  /** @nullable */
  iataCode: string | null;
  totalRatings: number;
  /** @nullable */
  averageRating: number | null;
}

export interface RatingsSummary {
  summary: AirlineRatingSummaryItem[];
  totalRatingsAllAirlines: number;
}

export type FetchFlightPassengersParams = {
date: string;
};

export type FetchFlightRatingsParams = {
date?: string;
};