# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: flights
    description: Pre-flight registration
  - name: ratings
    description: Anonymous flight ratings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /flights/register:
    post:
      operationId: registerFlight
      tags: [flights]
      summary: Register for an upcoming flight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightRegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightRegistrationResponse"
        "400":
          description: Bad request

  /flights/unregister:
    delete:
      operationId: unregisterFlight
      tags: [flights]
      summary: Cancel a flight registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnregisterFlightRequest"
      responses:
        "200":
          description: Successfully unregistered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnregisterResult"

  /flights/{flightNumber}/passengers:
    get:
      operationId: fetchFlightPassengers
      tags: [flights]
      summary: Get passenger count for a flight on a date
      parameters:
        - name: flightNumber
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Passenger information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightPassengers"
        "400":
          description: Bad request

  /flights/rating:
    post:
      operationId: submitRating
      tags: [ratings]
      summary: Submit or update an anonymous flight rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "200":
          description: Rating saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResult"
        "400":
          description: Bad request

  /flights/ratings/{flightNumber}:
    get:
      operationId: fetchFlightRatings
      tags: [ratings]
      summary: Get aggregated ratings for a specific flight
      parameters:
        - name: flightNumber
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Rating aggregate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightRatingAggregate"

  /ratings/airline/{iataCode}:
    get:
      operationId: getAirlineRatings
      tags: [ratings]
      summary: Get aggregated ratings for an airline by IATA code
      parameters:
        - name: iataCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Airline rating aggregate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirlineRatingAggregate"

  /ratings/summary:
    get:
      operationId: getRatingsSummary
      tags: [ratings]
      summary: Get rating summary for all airlines
      responses:
        "200":
          description: Ratings summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    FlightRegistrationInput:
      type: object
      properties:
        deviceId:
          type: string
        flightNumber:
          type: string
        flightDate:
          type: string
        passengerName:
          type: string
      required:
        - deviceId
        - flightNumber
        - flightDate

    FlightRegistrationResponse:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        flightNumber:
          type: string
        flightDate:
          type: string
        passengerName:
          type: ["string", "null"]
        registeredAt:
          type: string
      required:
        - id
        - deviceId
        - flightNumber
        - flightDate
        - registeredAt

    UnregisterFlightRequest:
      type: object
      properties:
        deviceId:
          type: string
        flightNumber:
          type: string
        flightDate:
          type: string
      required:
        - deviceId
        - flightNumber
        - flightDate

    UnregisterResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    FlightPassengers:
      type: object
      properties:
        flightNumber:
          type: string
        flightDate:
          type: string
        count:
          type: integer
        names:
          type: array
          items:
            type: string
      required:
        - flightNumber
        - flightDate
        - count
        - names

    RatingInput:
      type: object
      properties:
        deviceId:
          type: string
        flightNumber:
          type: string
        flightDate:
          type: string
        iataCode:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
      required:
        - deviceId
        - flightNumber
        - flightDate
        - rating

    RatingResult:
      type: object
      properties:
        success:
          type: boolean
        totalRatings:
          type: integer
        averageRating:
          type: ["number", "null"]
      required:
        - success
        - totalRatings
        - averageRating

    FlightRatingAggregate:
      type: object
      properties:
        flightNumber:
          type: string
        flightDate:
          type: ["string", "null"]
        totalRatings:
          type: integer
        averageRating:
          type: ["number", "null"]
      required:
        - flightNumber
        - flightDate
        - totalRatings
        - averageRating

    AirlineRatingAggregate:
      type: object
      properties:
        iataCode:
          type: string
        totalRatings:
          type: integer
        averageRating:
          type: ["number", "null"]
      required:
        - iataCode
        - totalRatings
        - averageRating

    AirlineRatingSummaryItem:
      type: object
      properties:
        iataCode:
          type: ["string", "null"]
        totalRatings:
          type: integer
        averageRating:
          type: ["number", "null"]
      required:
        - iataCode
        - totalRatings
        - averageRating

    RatingsSummary:
      type: object
      properties:
        summary:
          type: array
          items:
            $ref: "#/components/schemas/AirlineRatingSummaryItem"
        totalRatingsAllAirlines:
          type: integer
      required:
        - summary
        - totalRatingsAllAirlines