# lib/api-zod/src/generated/types/airlineRatingAggregate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AirlineRatingAggregate {
  iataCode: string;
  totalRatings: number;
  /** @nullable */
  averageRating: number | null;
}