# lib/api-zod/src/generated/types/airlineRatingSummaryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AirlineRatingSummaryItem {
  /** @nullable */
  iataCode: string | null;
  totalRatings: number;
  /** @nullable */
  averageRating: number | null;
}