# lib/api-zod/src/generated/types/flightRatingAggregate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlightRatingAggregate {
  flightNumber: string;
  /** @nullable */
  flightDate: string | null;
  totalRatings: number;
  /** @nullable */
  averageRating: number | null;
}