# lib/api-zod/src/generated/types/flightRegistrationResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlightRegistrationResponse {
  id: number;
  deviceId: string;
  flightNumber: string;
  flightDate: string;
  /** @nullable */
  passengerName?: string | null;
  registeredAt: string;
}