# lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  deviceId: string;
  flightNumber: string;
  flightDate: string;
  iataCode?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}