# lib/api-zod/src/generated/types/ratingsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AirlineRatingSummaryItem } from './airlineRatingSummaryItem';

export interface RatingsSummary {
  summary: AirlineRatingSummaryItem[];
  totalRatingsAllAirlines: number;
}